Optimize configurations, update readme

+++ include/RobotLeg.h
@@ -2,18 +2,17 @@
 #define ROBOTLEG_H
 
 #include <array>
-#include <cstddef>
 
 struct RobotLeg {
-    std::array<int, 3> channels;
+    std::array<uint8_t, 3> channels;
 
-    RobotLeg(int ch0, int ch1, int ch2) : channels{ch0, ch1, ch2} {}
+    RobotLeg(uint8_t ch0, uint8_t ch1, uint8_t ch2) : channels{ch0, ch1, ch2} {}
 
-    int& operator[](size_t index) {
+    uint8_t& operator[](size_t index) {
         return channels[index];
     }
 
-    const int& operator[](size_t index) const {
+    const uint8_t& operator[](size_t index) const {
         return channels[index];
     }
 };
